refector: 코인토스 애니매이션 코드 리팩토링
